feat(node): export `EngineArgs`
