+++ bin/reth/src/lib.rs
@@ -68,6 +68,58 @@ pub mod prometheus_exporter {
 /// change.
 pub mod args {
     pub use reth_node_core::args::*;
+
+    use clap::Args;
+    use reth_node_builder::engine_tree_config::{
+        DEFAULT_MEMORY_BLOCK_BUFFER_TARGET, DEFAULT_PERSISTENCE_THRESHOLD,
+    };
+
+    /// Parameters for configuring the engine
+    #[derive(Debug, Clone, Args, PartialEq, Eq)]
+    #[command(next_help_heading = "Engine")]
+    pub struct EngineArgs {
+        /// Enable the engine2 experimental features on reth binary
+        #[arg(long = "engine.experimental", default_value = "false")]
+        pub experimental: bool,
+
+        /// Configure persistence threshold for engine experimental.
+        #[arg(long = "engine.persistence-threshold", requires = "experimental", default_value_t = DEFAULT_PERSISTENCE_THRESHOLD)]
+        pub persistence_threshold: u64,
+
+        /// Configure the target number of blocks to keep in memory.
+        #[arg(long = "engine.memory-block-buffer-target", requires = "experimental", default_value_t = DEFAULT_MEMORY_BLOCK_BUFFER_TARGET)]
+        pub memory_block_buffer_target: u64,
+    }
+
+    impl Default for EngineArgs {
+        fn default() -> Self {
+            Self {
+                experimental: false,
+                persistence_threshold: DEFAULT_PERSISTENCE_THRESHOLD,
+                memory_block_buffer_target: DEFAULT_MEMORY_BLOCK_BUFFER_TARGET,
+            }
+        }
+    }
+
+    #[cfg(test)]
+    mod tests {
+        use super::*;
+        use clap::Parser;
+
+        /// A helper type to parse Args more easily
+        #[derive(Parser)]
+        struct CommandParser<T: Args> {
+            #[command(flatten)]
+            args: T,
+        }
+
+        #[test]
+        fn test_parse_engine_args() {
+            let default_args = EngineArgs::default();
+            let args = CommandParser::<EngineArgs>::parse_from(["reth"]).args;
+            assert_eq!(args, default_args);
+        }
+    }
 }
 
 /// Re-exported from `reth_node_core`, also to prevent a breaking change. See the comment on
